+++ src/integrations/terminal/TerminalRegistry.ts
@@ -5,6 +5,9 @@ export interface TerminalInfo {
 	busy: boolean
 	lastCommand: string
 	id: number
+	stream?: AsyncIterable<string>
+	running: boolean
+	streamClosed: boolean
 }
 
 // Although vscode.window.terminals provides a list of all open terminals, there's no way to know whether they're busy or not (exitStatus does not provide useful information for most commands). In order to prevent creating too many terminals, we need to keep track of terminals through the life of the extension, as well as session specific terminals for the life of a task (to get latest unretrieved output).
@@ -20,32 +23,59 @@ export class TerminalRegistry {
 			iconPath: new vscode.ThemeIcon("rocket"),
 			env: {
 				PAGER: "cat",
+
+				// VSCode bug#237208: Command output can be lost due to a race between completion
+				// sequences and consumers. Add 50ms delay via PROMPT_COMMAND to ensure the
+				// \x1b]633;D escape sequence arrives after command output is processed.
+				PROMPT_COMMAND: "sleep 0.050",
+
+				// VTE must be disabled because it prevents the prompt command above from executing
+				// See https://wiki.gnome.org/Apps/Terminal/VTE
+				VTE_VERSION: "0",
 			},
 		})
+
 		const newInfo: TerminalInfo = {
 			terminal,
 			busy: false,
 			lastCommand: "",
 			id: this.nextTerminalId++,
+			running: false,
+			streamClosed: false,
 		}
+
 		this.terminals.push(newInfo)
 		return newInfo
 	}
 
 	static getTerminal(id: number): TerminalInfo | undefined {
 		const terminalInfo = this.terminals.find((t) => t.id === id)
+
 		if (terminalInfo && this.isTerminalClosed(terminalInfo.terminal)) {
 			this.removeTerminal(id)
 			return undefined
 		}
+
 		return terminalInfo
 	}
 
 	static updateTerminal(id: number, updates: Partial<TerminalInfo>) {
 		const terminal = this.getTerminal(id)
+
 		if (terminal) {
 			Object.assign(terminal, updates)
 		}
+	}
+
+	static getTerminalInfoByTerminal(terminal: vscode.Terminal): TerminalInfo | undefined {
+		const terminalInfo = this.terminals.find((t) => t.terminal === terminal)
+
+		if (terminalInfo && this.isTerminalClosed(terminalInfo.terminal)) {
+			this.removeTerminal(terminalInfo.id)
+			return undefined
+		}
+
+		return terminalInfo
 	}
 
 	static removeTerminal(id: number) {